Add convenience scripts

+++ run.sh
@@ -3,7 +3,9 @@
 BASE_DIR=$(pwd)
 CONTAINER_NAME="lc0_base"
 
-docker container rm -f "$CONTAINER_NAME"
+if [ "$(docker ps -a | grep $CONTAINER_NAME)" ]; then
+  docker container rm -f "$CONTAINER_NAME"
+fi
 
 docker container run \
     -it \
